split constants outside version file
